% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-job-hub.R
\name{.ls_admin_job_hub}
\alias{.ls_admin_job_hub}
\title{Shared admin Shiny job push hub (one background worker per admin app process).}
\usage{
.ls_admin_job_hub
}
\description{
Shared admin Shiny job push hub (one background worker per admin app process).
}
\keyword{internal}
