% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-jobs-ui.R
\name{.ls_admin_jobs_tree_css}
\alias{.ls_admin_jobs_tree_css}
\title{CSS for admin job tree (shared look with LibeRation client)}
\usage{
.ls_admin_jobs_tree_css()
}
\description{
CSS for admin job tree (shared look with LibeRation client)
}
\keyword{internal}
