% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-jobs-ui.R
\name{.ls_admin_jobs_tree_js}
\alias{.ls_admin_jobs_tree_js}
\title{JavaScript handlers for admin job tree clicks}
\usage{
.ls_admin_jobs_tree_js()
}
\description{
JavaScript handlers for admin job tree clicks
}
\keyword{internal}
