% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-jobs-ui.R
\name{.ls_admin_jobs_tree_ui}
\alias{.ls_admin_jobs_tree_ui}
\title{Admin job tree list (matches LibeRation client Jobs tab style)}
\usage{
.ls_admin_jobs_tree_ui(df, selected_key, expanded_keys)
}
\description{
Admin job tree list (matches LibeRation client Jobs tab style)
}
\keyword{internal}
