% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto.R
\name{.ls_aead_encrypt_obj}
\alias{.ls_aead_encrypt_obj}
\title{AEAD-encrypt an arbitrary R object to a self-describing envelope list.}
\usage{
.ls_aead_encrypt_obj(obj, key)
}
\description{
AEAD-encrypt an arbitrary R object to a self-describing envelope list.
}
\keyword{internal}
