% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{.ls_dataset_list}
\alias{.ls_dataset_list}
\title{List registered datasets visible to a user (NULL = admin, sees all)}
\usage{
.ls_dataset_list(username = NULL)
}
\description{
List registered datasets visible to a user (NULL = admin, sees all)
}
\keyword{internal}
