% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto.R
\name{.ls_decrypt_from_file}
\alias{.ls_decrypt_from_file}
\title{Read and decrypt an envelope file with a key.}
\usage{
.ls_decrypt_from_file(key, path)
}
\description{
Read and decrypt an envelope file with a key.
}
\keyword{internal}
