% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto.R
\name{.ls_dek_generate}
\alias{.ls_dek_generate}
\title{Generate a fresh random data-encryption key (32 bytes).}
\usage{
.ls_dek_generate()
}
\description{
Generate a fresh random data-encryption key (32 bytes).
}
\keyword{internal}
