% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto.R
\name{.ls_encrypt_to_file}
\alias{.ls_encrypt_to_file}
\title{Encrypt an object with a key and persist the envelope to a file.}
\usage{
.ls_encrypt_to_file(obj, key, path)
}
\description{
Encrypt an object with a key and persist the envelope to a file.
}
\keyword{internal}
