% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{.ls_env}
\alias{.ls_env}
\title{Read env var, with optional legacy fallback name}
\usage{
.ls_env(name, legacy = NULL, unset = "")
}
\description{
Read env var, with optional legacy fallback name
}
\keyword{internal}
