% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{.ls_global_running_jobs}
\alias{.ls_global_running_jobs}
\title{Count jobs currently occupying a run slot across ALL users.}
\usage{
.ls_global_running_jobs()
}
\description{
Count jobs currently occupying a run slot across ALL users.
}
\keyword{internal}
