% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{.ls_is_dev_pkg_root}
\alias{.ls_is_dev_pkg_root}
\title{Detect whether a path is a development package source tree}
\usage{
.ls_is_dev_pkg_root(root)
}
\description{
Detect whether a path is a development package source tree
}
\keyword{internal}
