% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{.ls_job_has_result}
\alias{.ls_job_has_result}
\title{TRUE if a job produced a result, whether plaintext or encrypted.}
\usage{
.ls_job_has_result(job_path)
}
\description{
TRUE if a job produced a result, whether plaintext or encrypted.
}
\keyword{internal}
