% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{.ls_job_result_mtime}
\alias{.ls_job_result_mtime}
\title{Latest mtime among result files, or NA if none.}
\usage{
.ls_job_result_mtime(job_path)
}
\description{
Latest mtime among result files, or NA if none.
}
\keyword{internal}
