% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobs.R
\name{.ls_job_submit}
\alias{.ls_job_submit}
\title{Submit a job for a user}
\usage{
.ls_job_submit(username, limits, payload)
}
\arguments{
\item{username}{Authenticated user.}

\item{limits}{User limit list.}

\item{payload}{Parsed JSON body from API.}
}
\value{
Job metadata list.
}
\description{
Submit a job for a user
}
\keyword{internal}
