% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{.ls_job_worker_env}
\alias{.ls_job_worker_env}
\title{Worker package environment saved in each job directory}
\usage{
.ls_job_worker_env()
}
\description{
Worker package environment saved in each job directory
}
\keyword{internal}
