% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto.R
\name{.ls_new_salt_hex}
\alias{.ls_new_salt_hex}
\title{Generate a fresh random 32-byte salt (hex) for a user.}
\usage{
.ls_new_salt_hex()
}
\description{
Generate a fresh random 32-byte salt (hex) for a user.
}
\keyword{internal}
