% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto.R
\name{.ls_uk_forget}
\alias{.ls_uk_forget}
\title{Forget a cached user key (e.g. on token rotation).}
\usage{
.ls_uk_forget(username)
}
\description{
Forget a cached user key (e.g. on token rotation).
}
\keyword{internal}
