% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto.R
\name{.ls_uk_get}
\alias{.ls_uk_get}
\title{Retrieve a cached user key, or NULL if the owner has not authenticated.}
\usage{
.ls_uk_get(username)
}
\description{
Retrieve a cached user key, or NULL if the owner has not authenticated.
}
\keyword{internal}
