% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crypto.R
\name{.ls_uk_remember}
\alias{.ls_uk_remember}
\title{Remember a user's derived key for this server process (from an authed token).}
\usage{
.ls_uk_remember(username, token)
}
\description{
Remember a user's derived key for this server process (from an authed token).
}
\keyword{internal}
