% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{.ls_worker_load_pkg_root}
\alias{.ls_worker_load_pkg_root}
\title{Load LibeRation/LibeRtAD for a job worker process}
\usage{
.ls_worker_load_pkg_root(root, pkg, log_path = NULL)
}
\description{
Load LibeRation/LibeRtAD for a job worker process
}
\keyword{internal}
