% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker.R
\name{.ls_worker_log_snippet}
\alias{.ls_worker_log_snippet}
\title{Extract a useful error snippet from worker.log}
\usage{
.ls_worker_log_snippet(log_path, tail = 40L)
}
\description{
Extract a useful error snippet from worker.log
}
\keyword{internal}
