% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin-job-hub.R
\name{ls_admin_job_hub_register}
\alias{ls_admin_job_hub_register}
\title{Register an admin Shiny session with the shared job push hub.}
\usage{
ls_admin_job_hub_register(session)
}
\arguments{
\item{session}{Shiny session.}
}
\description{
Register an admin Shiny session with the shared job push hub.
}
