% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wire.R
\name{ls_job_to_wire}
\alias{ls_job_to_wire}
\title{Convert a LibeR job to the non-executable JSON wire contract}
\usage{
ls_job_to_wire(job)
}
\arguments{
\item{job}{A [ls_job()] object containing an `nm_model`.}
}
\value{
A JSON-compatible typed object.
}
\description{
Convert a LibeR job to the non-executable JSON wire contract
}
