% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job.R
\name{ls_queue_capabilities}
\alias{ls_queue_capabilities}
\title{Report queue and remote-worker contract capabilities}
\usage{
ls_queue_capabilities()
}
\description{
Report queue and remote-worker contract capabilities
}
