% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{ls_remote}
\alias{ls_remote}
\title{Connect to a LibeRties remote server}
\usage{
ls_remote(url, token, timeout = 60)
}
\arguments{
\item{url}{Server base URL.}

\item{token}{Bearer token issued by [ls_user_create()].}

\item{timeout}{HTTP timeout in seconds.}
}
\value{
A configured `LibeRRemote` client.
}
\description{
Connect to a LibeRties remote server
}
\examples{
remote <- ls_remote("https://liberties.example.org", "replace-with-token")
remote$url
}
