% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wire.R
\name{ls_result_decode}
\alias{ls_result_decode}
\title{Decode a remote result object}
\usage{
ls_result_decode(json, max_bytes = 500 * 1024^2)
}
\arguments{
\item{json}{UTF-8 result JSON.}

\item{max_bytes}{Maximum accepted result size.}
}
\description{
Decode a remote result object
}
