% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{ls_user_list}
\alias{ls_user_list}
\title{List server users without credential material}
\usage{
ls_user_list(root)
}
\arguments{
\item{root}{LibeRties server storage root.}
}
\description{
List server users without credential material
}
