% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{ls_user_remove}
\alias{ls_user_remove}
\title{Remove a user and optionally their sandbox}
\usage{
ls_user_remove(username, remove_sandbox = FALSE)
}
\arguments{
\item{username}{User to remove.}

\item{remove_sandbox}{If TRUE, delete the user's job directory.}
}
\description{
Remove a user and optionally their sandbox
}
