% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote.R
\name{LibeRRemote}
\alias{LibeRRemote}
\title{LibeRties remote-server client}
\description{
LibeRties remote-server client
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-LibeRRemote-initialize}{\code{LibeRRemote$new()}}
    \item \href{#method-LibeRRemote-authenticate}{\code{LibeRRemote$authenticate()}}
    \item \href{#method-LibeRRemote-submit}{\code{LibeRRemote$submit()}}
    \item \href{#method-LibeRRemote-list}{\code{LibeRRemote$list()}}
    \item \href{#method-LibeRRemote-status}{\code{LibeRRemote$status()}}
    \item \href{#method-LibeRRemote-result}{\code{LibeRRemote$result()}}
    \item \href{#method-LibeRRemote-logs}{\code{LibeRRemote$logs()}}
    \item \href{#method-LibeRRemote-cancel}{\code{LibeRRemote$cancel()}}
    \item \href{#method-LibeRRemote-clone}{\code{LibeRRemote$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LibeRRemote-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-LibeRRemote-initialize}{}}}
\subsection{\code{LibeRRemote$new()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LibeRRemote$new(url, token, timeout = 60)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LibeRRemote-authenticate"></a>}}
\if{latex}{\out{\hypertarget{method-LibeRRemote-authenticate}{}}}
\subsection{\code{LibeRRemote$authenticate()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LibeRRemote$authenticate()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LibeRRemote-submit"></a>}}
\if{latex}{\out{\hypertarget{method-LibeRRemote-submit}{}}}
\subsection{\code{LibeRRemote$submit()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LibeRRemote$submit(job)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LibeRRemote-list"></a>}}
\if{latex}{\out{\hypertarget{method-LibeRRemote-list}{}}}
\subsection{\code{LibeRRemote$list()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LibeRRemote$list()}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LibeRRemote-status"></a>}}
\if{latex}{\out{\hypertarget{method-LibeRRemote-status}{}}}
\subsection{\code{LibeRRemote$status()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LibeRRemote$status(id)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LibeRRemote-result"></a>}}
\if{latex}{\out{\hypertarget{method-LibeRRemote-result}{}}}
\subsection{\code{LibeRRemote$result()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LibeRRemote$result(id)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LibeRRemote-logs"></a>}}
\if{latex}{\out{\hypertarget{method-LibeRRemote-logs}{}}}
\subsection{\code{LibeRRemote$logs()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LibeRRemote$logs(id, stream = c("stdout", "stderr"))}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LibeRRemote-cancel"></a>}}
\if{latex}{\out{\hypertarget{method-LibeRRemote-cancel}{}}}
\subsection{\code{LibeRRemote$cancel()}}{
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LibeRRemote$cancel(id)}
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LibeRRemote-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LibeRRemote-clone}{}}}
\subsection{\code{LibeRRemote$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{LibeRRemote$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
