% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{ls_run_admin}
\alias{ls_run_admin}
\title{Launch the admin Shiny application}
\usage{
ls_run_admin(host = NULL, port = NULL)
}
\arguments{
\item{host}{Host (default from config \code{admin_host}).}

\item{port}{Port (default from config \code{admin_port}).}
}
\description{
Launch the admin Shiny application
}
