% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{ls_run_api}
\alias{ls_run_api}
\title{Start the scheduler Plumber API}
\usage{
ls_run_api(host = NULL, port = NULL)
}
\arguments{
\item{host}{Host to bind (default from config).}

\item{port}{Port (default from config).}
}
\description{
Start the scheduler Plumber API
}
